+++ udata_front/theme/gouvfr/datagouv-components/src/components/ResourceAccordion/ResourceAccordion.vue
@@ -52,6 +52,23 @@
             {{ $t('Visit') }}
           </a>
         </p>
+        <p class="fr-col-auto fr-ml-3v fr-m-0 z-2" v-else-if="ogcService">
+          <button
+            :id="resource.id + '-copy'"
+            :data-clipboard-text="resource.url"
+            class="fr-btn fr-btn--sm"
+          >
+            <span>
+                <OhVueIcon
+                  :height="16"
+                  :width="16"
+                  name="ri-file-copy-line"
+                  class="copy-icon fr-mr-2v"
+                />
+            </span> 
+            {{ t('Copy link') }}
+          </button>
+        </p>
         <p class="fr-col-auto fr-ml-3v fr-m-0" v-else>
           <a
             :href="resource.latest"
@@ -171,6 +188,9 @@ import TabPanel from "../Tabs/TabPanel.vue";
 import { trackEvent } from "../../helpers/matomo";
 import CopyButton from "../CopyButton/CopyButton.vue";
 import { getResourceFormatIconSvg } from "../../helpers/resources";
+import { OhVueIcon } from 'oh-vue-icons';
+
+const OGC_SERVICES_FORMATS = ['ogc:wfs', 'ogc:wms', 'wfs', 'wms'];
 
 const props = withDefaults(defineProps<{
   datasetId: string,
@@ -196,6 +216,8 @@ const hasPreview = computed(() => {
 })
 
 const format = computed(() => getResourceFormatIconSvg(props.resource) ? props.resource.format : t("File"))
+
+const ogcService = computed(() => OGC_SERVICES_FORMATS.includes(props.resource.format))
 
 const open = ref(props.expandedOnMount);
 const toggle = () => {