+++ udata_front/__init__.py
@@ -2,5 +2,5 @@
 udata customizations for data.gouv.fr
 '''
 
-__version__ = '6.0.6.dev'
+__version__ = '6.0.7.dev'
 __description__ = 'udata customizations for data.gouv.fr'